feat: Update Transaction model and repository for UUID support, enhance risk assessment logic, and modify SQL data insertion

+++ backend/src/main/java/com/workshop/backend/dto/TransactionDto.java
@@ -4,9 +4,9 @@ import lombok.Data;
 
 @Data
 public class TransactionDto {
-    private String cc_number;
-    private float amount;
-    private String category;
-    private float latitude;
-    private float longitude;
+    private String cc_number;  // Credit card number
+    private Double amount;     // Transaction amount
+    private String category;   // Transaction category
+    private Double latitude;   // Location latitude
+    private Double longitude;  // Location longitude
 }
